+++ src/main/kotlin/presentation/features/GetIraqiMealsUI.kt
@@ -1,4 +1,4 @@
-package org.example.presentation
+package org.example.presentation.features
 
 import org.example.logic.usecases.GetIraqiMealsUseCase
