+++ src/main/kotlin/presentation/features/SearchMealByNameUI.kt
@@ -1,8 +1,7 @@
-package presentation
+package org.example.presentation.features
 
 import org.example.logic.model.FoodChangeModeResults
 import org.example.logic.usecases.GetMealByNameUseCase
-
 
 class SearchMealByNameUI(private val getMealByNameUseCase: GetMealByNameUseCase) {
