+++ src/main/kotlin/presentation/features/SuggestSweetNoEggsUI.kt
@@ -1,4 +1,4 @@
-package presentation;
+package org.example.presentation.features
 
 import org.example.logic.usecases.GetDessertsWithNoEggs
